Model's PSF normalized, faster psf profile
By always pre-normalizing the PSF given by the user we gain in two ways.

Firstly, and most importantly, the "psf" profile has always worked
implicitly on a normalized PSF. It previously couldn't assume the
Model's PSF was normalized, and therefore it had to calculate the total
flux of the PSF, and take it into account when doing the final scaling
of the image by the requested magnitude. The way we calculated the total
flux was *very* awkward to begin with (it was implemented before the
Image class days), and on top of that the final scaling was performed
over the *whole* image, which was inefficient (because the problem
suddenly scaled with the size of the image and not the PSF's). Having a
pre-normalized PSF makes this code much simpler, as there is no
accumulation required, and the scaling can be done directly when
assigning values to the output pixels.

The other place where the PSF is used is during convolution. We were
previously always normalizing the PSF and producing a new Image, and
this required an extra memory allocation. This need is gone now, making
this an extra, smaller improvement.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>