Avoiding memory allocation on FITS dumping
We can avoid a potentially big and expensive memory allocation when
doing the conversion from little to big ending needed for FITS data
dumping. Instead, we loop through the pixels, convert them individually,
and write them into the file immediately after conversion. OS caching
saves us from having a bad writing performance most probably, and anyway
it's not really an issue for us as the tool is mostly exploratory and is
not meant to be *that* performant.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>