Minor improvements to image-related classes
We previously offered single-element constructors with a default value,
but that at the end initialized the underlying vector with an explicit
size of 0. I'm not exactly sure whether this may or may not cause a
memory allocation, but better safe than sorry!

I'm also making the Box::empty() method const, there was no reason for
it not to be const.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>