Profiles add to pixels instead of setting values
Until now we have always given each profile a "clean slate": an image
that is known to have only zeroes in it. This has given profile classes
the freedom to *set* values on each pixel without caring about their
previous content, and to perform "whole image" operations if required.
This however has a few costs: firstly, an extra memory allocation is
required for the scratch area that is given to each profile for
evaluation. Secondly, the scratch area must be reset to contain only
zeroes before evaluating the next profile. Finally, the scratch area
needs to be added into the final image after each evaluation. The final
two costs can be offset somewhat if we knew the drawing boundaries of
each profile (which would lead to more complex code), but until now we
have been doing this operation image-wide, which is unnecessary for
profiles like "null" or "psf", and for many parameter combinations of
the different radial profiles.

With the latest changes to the PSF profile, no profile requires to
perform image-wide operations anymore. Thus, changing profiles to *add*
to a single image surface instead of giving them a clean surface is now
possible. This commit changes the behaviour of all profiles to be
additive instead of overwriting the contents of the image's pixels. The
Model still needs to keep track of possibly (but not necessarily) two
different Image objects which are giving to each profile depending on
whether the latter needs convolution or not. With this change there is
no need for a separate scratch area anymore, saving us the extra memory
allocation, and there is no need to perform any further image addition,
cutting down that cost as well.

The only test that needed adjustment was the one where we compared an
image coming from a Model with many profiles on the one hand, and the
"manual" addition of images coming from many Models, each with one of
those profiles on the other. We were previously testing for an *exact*
match which is not the case anymore. I believe this is due to a
difference on how addition is performed: in the case of a Model with
many profiles we previously performed image-wide addition which resulted
(in native builds) in assembly code using the widest registers available
and performing addition in those, whereas now we perform addition on a
pixel-by-pixel basis, and those registers, and their associated
operations, are not used anymore. I am assuming the addition operations
on two different register types *might* be able to yield slightly
different results, which would then explain the difference in the
images.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>