Complete revamp of project structure
The old project structure with the source files separated from the
include files and from the tests had become a bit difficult to maintain,
and prevented us from doing some interesting things with cmake down the
line.

The new structure makes things easier. At the top we have the src/
subdirectory, which contains the profit/ and apps/ subdirectories. The
former has all the bits necessary to build libprofit.so, while the
latter has the sources for profit-cli. The tests that are
profit-cli-specific are added in src/apps, while the rest are added in
src/profit. In terms of code *nothing* has changed though, only the
files' placement on the directory structure.

On top of this, more fine-grained control over inclusion of directories and
library linking is done now, with both happening at the target level.
Most directory inclusion and linkage does not need to be public
actually, so they are declared as PRIVATE instead. Doing this also
exposed how much we were relying on the global behavior, as both the
profit-cli and the test_util executables used GSL separately, and
therefore both targets needed modification. To avoid too many
GSL-related operations I created a GSL::gsl target when one is not
available, so executables (and libprofit itself) link to it.

All these changes are meant to make it easier to export libprofit's
targets and provide (and install) cmake package configuration files
which downstream projects can use.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>