Speeding up coordinate conversion
The previous code doing coordinate conversion was previously doing a
division of the resulting yprof coordinate over axrat. We know fp
divisions are way slower than the rest of the operations, and can easily
skip this one by pre-calculating the division of cos_ang and sin_ang
over axrat, and using *that* when calculating yprof.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>