Avoiding OpenMP setup when not required
This removes small overheads due to OpenMP setups when the code has been
compiled with OpenMP support but users use a single thread. In such
cases we definitely don't want to spend time setting up infrastructure
we will not end up using. This amounts to ~40, 50 [us] in a native,
release build in my machine, which is ~5% of some non-boxy sersic
profiles.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>