Update MacOS Travis job specifications
MacOS 10.15 is now the lowest version supported by homebrew; therefore
to avoid any issues we are greatly simplifying our Travis infrastructure
to only test against this one version of MacOS.

As part of this upgrade I'm also taking the opportunity to avoid
upgrading homebrew. Skipping these upgrades yields a faster Travis CI
execution, as well as a more stable platform on which we test (because
dependencies won't change).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>