Fix OpenCL program building from cache
When building OpenCL programs from the cache there was a small catch: we
were building the programs for all devices in the OpenCL Context instead
of for the device the program had been compiled for. This wasn't
apparent in many installations, but there had been already been hints to
this problem in some MacOS installations in the past, plus now Travis CI
pointing out the same error.

Building the cached program for a given device means we can reuse the
bit of code that does the building and issues an informative error
message with the build log back to the user.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>