Create error message only when required
In assert_images_relative_delta we construct an error message to display
when pixels between images differ from more than a given tolerance. This
error message was always created though, even when there was no error,
which lead to longer test execution times.

This commit moves the creation of the error message to be executed only
when the error condition is true; this means we can't use the more
convenient TSM_ASSERT_LESS_THAN_EQUALS macro but a more manual condition
check, but on the other hand it greatly reduces the walltime needed by
our tests.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>