Switch to std::fill for improved performance
Using vector.assign to zero out a vector's content is not as efficient
as I thought it would. It seems to run at ~3 GB/s or so, while a memset
operation should be much faster than that.

In this commit I'm changing the code to use std::fill() instead.
Switching to this function means the code compiles out into a call to
memset, which is the fastest way (within reason) we can achieve this
operation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>