Re-build OpenCL programs if load from cache fails
Even after fixing the rebuilding of programs from binaries, which I
thought was what was causing issues in MacOS Mojave+ with Intel, and
after reading more online, it became clear that this functionality is
broken in this particular platform combination. Like this there might be
others as well, so we better take care of it more generically.

This commit changes the logic of building a program such that when
building from cache fails we build from source and cache the result
again. This should make buggy build-from-binaries platforms still work
at the cost of always rebuilding the program from source, which is
anyway the only realistic solution.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>