Add method to query drawing dimensions
If we want to be able to pass pre-allocated Image objects to a Model
instance at evaluation time, we need to know the dimensions the Image
should have.

This commit adds a new method to the Model class to query the drawing
dimensions. The results is valid for the given Model configuration; as
soon as any of its parameter changes the method should be queried again
to get a valid result.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>